Create rusty-hog-scanner subcrate
Separate scanner into a different subcrate using a workspace. This way the scanner can be easily used in other projects without having to import all the other functionality such as git or s3 secrets reader.
This also makes it possible to use the rusty-hog-scanner in wasm modules.
