+++ crates/rusty-hog-scanner/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "rusty_hog_scanner"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = "1.0"
+serde_json = "1.0"
+serde_derive = "^1"
+clap = "2"
+simple_logger = "1.11"
+simple-error = "0.2"
+anyhow = "1.0"
+log = "0.4"
+base64 = "0.13"
+regex = "1"
+hex = "0.4"
