+++ crates/rusty-hog-scanner/src/default_allowlist.json
