+++ crates/rusty-hog-scanner/src/default_rules.json
