+++ src/aws_scanning.rs
@@ -50,7 +50,6 @@
 //! assert_eq!(results.len(), 0);
 //! ```
 
-use crate::SecretScanner;
 use encoding::all::ASCII;
 use encoding::{DecoderTrap, Encoding};
 use log::{self, error, trace};
@@ -58,6 +57,7 @@ use s3::bucket::Bucket;
 use serde_derive::{Deserialize, Serialize};
 use simple_error::SimpleError;
 use std::str;
+use rusty_hog_scanner::SecretScanner;
 
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Hash, Clone, Default)]
 /// `serde_json` object that represents a single found secret - finding
