+++ src/bin/berkshire_hog.rs
@@ -39,7 +39,7 @@ use std::str;
 use url::Url;
 
 use rusty_hogs::aws_scanning::{S3Finding, S3Scanner};
-use rusty_hogs::{SecretScanner, SecretScannerBuilder};
+use rusty_hog_scanner::{SecretScanner, SecretScannerBuilder};
 use std::collections::HashSet;
 
 /// Main entry function that uses the [clap crate](https://docs.rs/clap/2.33.0/clap/)
