+++ src/bin/berkshire_hog_lambda.rs
@@ -19,7 +19,7 @@ extern crate s3;
 use lambda_runtime::{handler_fn, Context, Error};
 use log::{self, warn, LevelFilter};
 use rusty_hogs::aws_scanning::{S3Finding, S3Scanner};
-use rusty_hogs::SecretScannerBuilder;
+use rusty_hog_scanner::SecretScannerBuilder;
 use s3::bucket::Bucket;
 use s3::creds::Credentials;
 use s3::region::Region;
