+++ src/bin/essex_hog.rs
@@ -44,8 +44,8 @@ use hyper::http::Request;
 use hyper::http::StatusCode;
 use hyper::{client, Body, Client};
 use log::{self, debug, error, info};
-use rusty_hogs::SecretScannerBuilder;
-use rusty_hogs::{RustyHogMatch, SecretScanner};
+use rusty_hog_scanner::SecretScannerBuilder;
+use rusty_hog_scanner::{RustyHogMatch, SecretScanner};
 use serde_derive::{Deserialize, Serialize};
 use serde_json::{Map, Value};
 use simple_error::SimpleError;
